+++ vclib/src/commonTest/kotlin/at/asitplus/wallet/lib/agent/DummyCredentialDataProvider.kt
@@ -16,42 +16,34 @@ class DummyCredentialDataProvider(
 
     private val defaultLifetime = 1.minutes
 
-    override fun getCredentialWithType(
-        subjectId: String,
-        subjectPublicKey: CryptoPublicKey?,
-        attributeTypes: Collection<String>,
+    override fun getCredential(
+        subjectPublicKey: CryptoPublicKey,
+        credentialScheme: ConstantIndex.CredentialScheme,
         representation: ConstantIndex.CredentialRepresentation
     ): KmmResult<List<CredentialToBeIssued>> {
-        val attributeType = ConstantIndex.AtomicAttribute2023.vcType
-        if (!attributeTypes.contains(attributeType)) {
-            return KmmResult.failure(UnsupportedOperationException("no data"))
-        }
         val expiration = clock.now() + defaultLifetime
         val claims = listOf(
             ClaimToBeIssued("given-name", "Susanne"),
             ClaimToBeIssued("family-name", "Meier"),
             ClaimToBeIssued("date-of-birth", "1990-01-01"),
         )
+        val subjectId = subjectPublicKey.toJsonWebKey().identifier
         val credentials = when (representation) {
             ConstantIndex.CredentialRepresentation.SD_JWT -> listOf(
                 CredentialToBeIssued.VcSd(
-                    subjectId = subjectId,
                     claims = claims,
                     expiration = expiration,
-                    scheme = ConstantIndex.AtomicAttribute2023,
                 )
             )
 
             ConstantIndex.CredentialRepresentation.PLAIN_JWT -> claims.map { claim ->
-                CredentialToBeIssued.Vc(
+                CredentialToBeIssued.VcJwt(
                     subject = AtomicAttribute2023(subjectId, claim.name, claim.value),
                     expiration = expiration,
-                    scheme = ConstantIndex.AtomicAttribute2023
                 )
-            } + CredentialToBeIssued.Vc(
+            } + CredentialToBeIssued.VcJwt(
                 subject = AtomicAttribute2023(subjectId, "picture", "foo"),
                 expiration = expiration,
-                scheme = ConstantIndex.AtomicAttribute2023,
                 attachments = listOf(Issuer.Attachment("picture", "image/webp", byteArrayOf(32)))
             )
 
@@ -60,9 +52,7 @@ class DummyCredentialDataProvider(
                     issuerSignedItems = claims.mapIndexed { index, claim ->
                         buildIssuerSignedItem(claim.name, claim.value, index.toUInt())
                     },
-                    subjectPublicKey = subjectPublicKey!!.toCoseKey(),
                     expiration = expiration,
-                    scheme = ConstantIndex.AtomicAttribute2023
                 )
             )
         }